remove pipeline state!!!!!!
