+++ AGENTS.md
@@ -44,7 +44,7 @@ Top-level packages of the library.
 - `lr_scheduler/` - learning-rate schedules, including `piecewise` (composable schedules).
 - `tracker/` - experiment tracking integrations (`provider`, e.g. WandB, Aim).
 - `kernel/` - custom kernels: `cce`, `flash_attn`, `gmm`, `moe`, `normalization`, `stochastic`, `swiglu`, `general`.
-- `internals/` - engine internals: `pipeline_state`, `grad_sync`, `grad_norm`, `metric_collector`, `determinism`, `profiling`, `state`.
+- `internals/` - engine internals: `grad_sync`, `grad_norm`, `metric_collector`, `determinism`, `profiling`, `state`.
 
 ## Working agreements for agents
