+++ zensical.toml
@@ -90,7 +90,6 @@ nav = [
         "internals/grad_norm.md",
         "internals/grad_sync.md",
         "internals/metric_collector.md",
-        "internals/pipeline_state.md",
         "internals/pipelining.md",
         "internals/profiling.md",
         "internals/tracker_integration.md",
